extract some function for embed usage
